Project 4 Ghostbusters/ai_dqn_project/requirements.txt
gym
tensorflow
opencv-python
Pillow
